Wisetrack/Final/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
requests
pytz
pandas
numpy
flask